Syndicate/Backend/requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.14
django-cors-headers>=4.3
python-dotenv>=1.0
openai>=1.40
pypdf>=4.0
python-docx>=1.1
